Reorganize ovarall msg structs
